Minor bug fixes

+++ nornir_buildmanager/operations/registration.py
@@ -4,16 +4,17 @@ Created on Jun 4, 2012
 @author: Jamesan
 '''
 
-from nornir_buildmanager import *
-from nornir_imageregistration.io import  *
-from nornir_shared.processoutputinterceptor import ProcessOutputInterceptor, ProgressOutputInterceptor
-import subprocess
-import shutil
-from nornir_shared import *
-import os
 import logging
-from nornir_buildmanager.validation import transforms
+import os
+import shutil
+import subprocess
 
+from nornir_buildmanager import *
+from nornir_buildmanager.validation import transforms
+from nornir_imageregistration.files import *
+from nornir_shared import *
+from nornir_shared.processoutputinterceptor import ProcessOutputInterceptor, \
+    ProgressOutputInterceptor
 
 
 def TranslateTransform(Parameters, TransformNode, LevelNode, Logger, **kwargs):
@@ -75,7 +76,17 @@ def TranslateTransform(Parameters, TransformNode, LevelNode, Logger, **kwargs):
             NewP = subprocess.Popen(cmd + " && exit", shell=True, stdout=subprocess.PIPE)
             ProcessOutputInterceptor.Intercept(ProgressOutputInterceptor(NewP))
             OutputTransformNode.cmd = cmd
-            SaveRequired = True
+
+            if os.path.exists(OutputTransformNode.FullPath):
+                stats = os.stat(OutputTransformNode.FullPath)
+                if stats.st_size == 0:
+                    os.remove(OutputTransformNode.FullPath)
+                    errmsg = "ir-refine-translate output zero size translate file.  Output deleted: " + OutputTransformNode.FullPath
+                    Logger.error(errmsg)
+
+                    # raise Exception(errmsg)
+
+            SaveRequired = os.path.exists(OutputTransformNode.FullPath)
         finally:
             if os.path.exists(mosaicFullPath):
                 os.remove(mosaicFullPath)
@@ -161,8 +172,6 @@ def CompressTransforms(Parameters, TransformNode, **kwargs):
         prettyoutput.Log("Input transform file not found: " + TransformNode.FullPath)
         return
 
-
-
     InputFileFullPath = TransformNode.FullPath
     MosaicBaseName = os.path.basename(InputFileFullPath)
     TempMosaicFilename = 'Temp_' + MosaicBaseName
@@ -172,10 +181,10 @@ def CompressTransforms(Parameters, TransformNode, **kwargs):
     mFile.CompressTransforms()
     mFile.Save(TempMosaicFileFullPath)
 
-    shutil.move(TempMosaicFileFullPath, InputFileFullPath)
-
-    TransformNode.Checksum = mFile.Checksum
-    TransformNode.attrib['Compressed'] = 'True'
+    if os.path.exists(TempMosaicFileFullPath):
+        shutil.move(TempMosaicFileFullPath, InputFileFullPath)
+        TransformNode.Checksum = mFile.Checksum
+        TransformNode.attrib['Compressed'] = 'True'
 
     return TransformNode.Parent
